pyproject.toml
[project]
name = "swimba-pipeline"
version = "0.1.0"
description = "Tools for generating swimba runs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camb>=1.6.4",
    "h5py>=3.15.1",
    "numpy>=2.3.4",
    "pyyaml>=6.0.3",
    "swiftemulator>=1.3.1",
    "pygadgetreader",
]

[project.scripts]
make_run = "swimba_pipeline.make_run:main"
convert_ic = "swimba_pipeline.convert_ic:main"
make_1P = "swimba_pipeline.make_1P:main"
make_LH = "swimba_pipeline.make_LH:main"
make_SB = "swimba_pipeline.make_SB:main"
make_CV = "swimba_pipeline.make_CV:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
pygadgetreader = { git = "https://github.com/dnarayanan/pygadgetreader" }
